Move the edit function to its own file
